Fixed a bug with initialising the selected base index.
The bug was to do with -1 being used to indicate that the base index was not set. However, this was ambiguous because -1 can be a real coordinate. This was the cause of the problems seen in RT199350, where Blixem was showing an incorrect coordinate for the selected base index in protein mode, if the user had not previously selected an index. Also in this change are a small fix to some help page text and the removal of a (currently) unused menu item.