Added exon view for vertical sequence in Dotter (WIP)
Work in progress. Added widgets to show exons for the vertical sequence in Dotter as well as the horizontal sequence. Fixed a bug so that an exon view only shows exons for the relevant sequence. Fixed a bug with calling Dotter on the reference sequence vs itself where it was passing the incorrect range for the second sequence. Reorganised code so that Dotter can call the finaliseBlxSequences function to construct any missing transcript data.