Fixed a performance issue with middle-dragging on the big picture
Previously the code was pushing the cached pixmap to screen for every mouse-motion and drawing the preview box on top. This was causing very slow re-draw on some deskpros (I think this is a GDK issue on Lenny deskpros with an external graphics card). I've changed it to just draw the preview box using XOR. (Unfortunately this means the colours are not so nice while dragging.)