Fixed a bug where out-of-range exons were being displayed
The problem was that the cell renderer was not filtering exons by display range. The bug was only apparent with the 'squash matches' option enabled because otherwise there is only one exon per row and the row gets filtered out by the tree if that exon is out of range.