Fixed some bugs with dottering the reverse strand vs itself
The auto coords were being calculated incorrectly for protein mode (needed to convert to DNA coords). The wrong strand was being passed to dotter (we should always pass the forward strand). The vertical sequence scale was not being reversed - I added some options to dotter to be able to specify whether each scale should be reversed.