Idential short-read alignments are now treated as the same sequence
Previously grouping was done by sequence name, but the name is not so relevant for short reads so we check whether the alignment is identical instead and, if so, treat them as the same sequence. This means that duplicates are filtered out when the 'squash matches' option is turned on.