Improvements to the toolbar and menu
Toolbar and menu are now on the same line to save space. Toolbar items are now for the same actions as the original 'one-click' buttons from the old belvu. Also consolodated the 'edit residue colors' and 'edit conservation colors' menu items into a single menu item that affects the currently-active scheme.