Fix a problem with font width highlighting in belvu - BUG FIX
This only seems reproducible on mac. The problem is most obvious with a wide sequence because the inaccuracy is cumulative from left to right.

There is still a problem with certain font sizes. As you increment the font size it goes 8(ok) -9-10-11-12- 12(ok) -13-14-15-16- 16(ok)
so it is only accurate for sizes 8, 12, 16 etc.